Bootstrap Nested Linear Mixed-Effects Models

Perform various bootstrap process for nested linear mixed effects (LMEs) models including: parametric, residual, cases, CGR, and REB bootstraps.

# S3 method for lmerMod
bootstrap(model, .f, type, B, resample, reb_type)

# S3 method for lme
bootstrap(model, .f, type, B, resample, reb_type)

bootstrap(model, .f, type, B, resample = NULL, reb_type = NULL)

Arguments

model
The model object you wish to bootstrap.

.f
A function returning the statistic(s) of interest.

type
A character string indicating the type of bootstrap that is being requested. Possible values are "parametric", "residual", "case", "cgr", or "reb" (random effect block bootstrap).

B
The number of bootstrap resamples.

resample
A logical vector specifying whether each level of the model should be resampled in the cases bootstrap. The levels should be specified from the highest level (largest cluster) of the hierarchy to the lowest (observation-level); for example for students within a school, specify the school level first, then the student level.

reb_type
Specification of what random effect block bootstrap version to implement. Possible values are 0, 1 or 2.

Value

The returned value is an object of class "lmeresamp".

Details

All of the below methods have been implemented for nested linear mixed-effects models fit by lmer (i.e., an lmerMod object) and lme (i.e., an lmerMod object). Details of the bootstrap procedures can be found in the help file for that specific function.

References

Carpenter, J. R., Goldstein, H. and Rasbash, J. (2003) A novel bootstrap procedure for assessing the relationship between class size and achievement. Journal of the Royal Statistical Society. Series C (Applied Statistics), 52, 431--443.

Chambers, R. and Chandra, H. (2013) A random effect block bootstrap for clustered data. Journal of Computational and Graphical Statistics, 22, 452--470.

Morris, J. S. (2002) The BLUPs are not "best" when it comes to bootstrapping. Statistics and Probability Letters, 56, 425--430.

Van der Leeden, R., Meijer, E. and Busing F. M. (2008) Resampling multilevel models. In J. de Leeuw and E. Meijer, editors, Handbook of Multilevel Analysis, pages 401--433. New York: Springer.

Bates, D., Maechler, M., Bolker, W., Walker, S. (2015). Fitting Linear Mixed-Effects Models Using lme4. Journal of Statistical Software, 67, 1--48. doi:10.18637/jss.v067.i01.

See also

Examples

library(lme4)

#> Loading required package: Matrix

vcmodA <- lmer(mathAge11 ~ mathAge8 + gender + class + (1 | school), data = jsp728)

## you can write your own function to return stats, or use something like 'fixef'
mySumm <- function(.) {
  s <- getME(., "sigma")
    c(beta = getME(., "beta"), sigma = s, sig01 = unname(s * getME(., "theta")))
}

## running a parametric bootstrap 
set.seed(1234)
boo1 <- bootstrap(model = vcmodA, .f = mySumm, type = "parametric", B = 100)

if (FALSE) {
## running a cases bootstrap - only resampling the schools
boo2 <- bootstrap(model = vcmodA, .f = mySumm, type = "case", B = 100, resample = c(TRUE, FALSE))

## running a cases bootstrap - resampling the schools and students within the school
boo2 <- bootstrap(model = vcmodA, .f = mySumm, type = "case", B = 100, resample = c(TRUE, FALSE))

## running a semi-parametric bootstrap
boo3 <- bootstrap(model = vcmodA, .f = mySumm, type = "cgr", B = 100)

## running a residual bootstrap
boo4 <- bootstrap(model = vcmodA, .f = mySumm, type = "residual", B = 100)

## running an REB0 bootstrap
boo5 <- bootstrap(model = vcmodA, .f = mySumm, type = "reb", B = 100, reb_typ = 0)
}

## to print results in a formatted way
print(boo1)

#> Bootstrap type: parametric 
#> 
#> Number of resamples: 100 
#> 
#>         observed   rep.mean         se          bias
#> beta1 14.1577509 14.0158661 0.65832020 -0.1418848720
#> beta2  0.6388895  0.6429228 0.02436914  0.0040333457
#> beta3 -0.3571922 -0.2909488 0.36701673  0.0662434573
#> beta4  0.7200815  0.7242379 0.39057968  0.0041564090
#> sigma  4.4416279  4.4414617 0.11345797 -0.0001661894
#> sig01  1.8199553  1.7909477 0.24175360 -0.0290075852