CGR Bootstrap for Nested LMEs

Generate semi-parametric bootstrap replicates of a statistic for a nested linear mixed-effects model.

# S3 method for lmerMod
cgr_bootstrap(model, .f, B)

# S3 method for lme
cgr_bootstrap(model, .f, B)

cgr_bootstrap(model, .f, B)

Arguments

model
The model object you wish to bootstrap.

.f
A function returning the statistic(s) of interest.

B
The number of bootstrap resamples.

Value

The returned value is an object of class "lmeresamp".

Details

The semi-parametric bootstrap algorithm implemented was outlined by Carpenter, Goldstein and Rasbash (2003). The algorithm is outlined below:

Obtain the parameter estimates from the fitted model and calculate the estimated error terms and EBLUPs.

Rescale the error terms and EBLUPs so that the empirical variance of these quantities is equal to estimated variance components from the model.

Sample independently with replacement from the rescaled estimated error terms and rescaled EBLUPs.

Obtain bootstrap samples by combining the samples via the fitted model equation.

Refit the model and extract the statistic(s) of interest.

Repeat steps 3-5 B times.

References

Carpenter, J. R., Goldstein, H. and Rasbash, J. (2003) A novel bootstrap procedure for assessing the relationship between class size and achievement. Journal of the Royal Statistical Society. Series C (Applied Statistics), 52, 431--443.

See also