Calculate summary statistics for lmeresamp objects

Calculate basic summary statistics such as the mean, standard error, and bias of the bootstrap replicates.

# S3 method for lmeresamp
summary(object, ...)

Arguments

object
The lmeresamp object to be summarized.

...
not used

Details

If the bootstrap statistics are stored in a vector (as opposed to a data frame or tibble), then summary statistics will be calculated. The printed data frame will include the name of the term (if applicable), the observed value (observed), the mean of the bootstrap replicated (rep.mean), the standard error (se), and the bootstrap bias estimate (bias). In addition, the number of resamples will be printed. If any messages, warnings, or errors were generated during the bootstrap procedure, they will be summarized below, and you should check the message, warning, and error elements of the lmeresamp object to investigate further.