Wild Bootstrap for LME models

Generate wild bootstrap replicates of a statistic for a linear mixed-effects model.

# S3 method for lmerMod
wild_bootstrap(model, .f, B, hccme = c("hc2", "hc3"), aux.dist = c("f1", "f2"))

# S3 method for lme
wild_bootstrap(model, .f, B, hccme = c("hc2", "hc3"), aux.dist = c("f1", "f2"))

wild_bootstrap(model, .f, B, hccme, aux.dist)

Arguments

model
The model object you wish to bootstrap.

.f
A function returning the statistic(s) of interest.

B
The number of bootstrap resamples.

hccme
either "hc2" or "hc3", indicating which heteroscedasticity consistent covariance matrix estimator to use.

aux.dist
either "f1" or "f2" indicating which auxiliary distribution to draw the errors from

Value

The returned value is an object of class "lmeresamp".

Details

The wild bootstrap algorithm for LMEs implemented here was outlined by Modugno & Giannerini (2015). The algorithm is outlined below:

Draw a random sample equal to the number of groups (clusters) from an auxillary distribution with mean zero and unit variance. Denote these as \(w_1, \ldots, w_g\).

Calculate the selected heteroscedasticity consistent matrix estimator for the marginal residuals, \(\tilde{v}_i\)

Generate bootstrap responses using the fitted equation: \(y_i^* = X_i \beta + \tilde{v}_i w_j\)

Refit the model and extract the statistic(s) of interest.

Repeat steps 2-4 B times.

References

Modugno, L., & Giannerini, S. (2015). The Wild Bootstrap for Multilevel Models. Communications in Statistics -- Theory and Methods, 44(22), 4812--4825.

See also